Add utility to generate title from slug
