+++ site/src/components/navigation/navigation.tsx
@@ -1,6 +1,7 @@
 import { graphql, Link, useStaticQuery } from "gatsby"
 import * as React from "react"
 import "./navigation.css"
+import createTitle from "../../utils/title"
 
 const Navigation: React.FC = () => {
   const { allMarkdownRemark } = useStaticQuery(graphql`
@@ -17,11 +18,12 @@ const Navigation: React.FC = () => {
     }
   `)
 
+
   return (
     <aside className="navigation">
       <h1><Link to="/">Git Guide</Link></h1>
       <ul>
-        {allMarkdownRemark.edges.map(({node}) => <li><Link to={node.fields.slug}>{node.fields.slug}</Link></li>)}
+        {allMarkdownRemark.edges.map(({node}) => <li><Link to={node.fields.slug}>{createTitle(node.fields.slug)}</Link></li>)}
       </ul>
     </aside>
   )
